refactor(api-test): remove @gbfm/api-test package, consolidate into scripts
