+++ apps/vps/package.json
@@ -72,7 +72,6 @@
   },
   "devDependencies": {
     "@effect/language-service": "0.86.4",
-    "@gbfm/api-test": "workspace:*",
     "@testcontainers/postgresql": "^11.14.0",
     "@types/aws-lambda": "8.10.161",
     "@types/bun": "1.3.13",
