--- packages/api-test/package.json
@@ -1,22 +1,0 @@
-{
-  "name": "@gbfm/api-test",
-  "version": "0.0.1",
-  "private": true,
-  "type": "module",
-  "exports": {
-    "./health": "./src/health.ts"
-  },
-  "scripts": {
-    "typecheck": "tsgo --noEmit",
-    "test": "vitest run"
-  },
-  "dependencies": {
-    "@gbfm/api": "workspace:*",
-    "effect": "4.0.0-beta.93"
-  },
-  "devDependencies": {
-    "@types/bun": "1.3.13",
-    "typescript": "6.0.3",
-    "vitest": "4.1.8"
-  }
-}
