--- packages/api-test/tsconfig.json
@@ -1,25 +1,0 @@
-{
-  "extends": "@tsconfig/bun/tsconfig.json",
-  "compilerOptions": {
-    "target": "ES2022",
-    "module": "preserve",
-    "moduleResolution": "bundler",
-    "moduleDetection": "force",
-    "isolatedModules": true,
-    "skipLibCheck": true,
-    "verbatimModuleSyntax": true,
-    "rewriteRelativeImportExtensions": true,
-    "strict": true,
-    "exactOptionalPropertyTypes": true,
-    "noUnusedLocals": true,
-    "noImplicitOverride": true,
-    "declarationMap": true,
-    "sourceMap": true,
-    "composite": true,
-    "noEmit": true,
-    "types": ["@types/bun"],
-    "plugins": [{ "name": "@effect/language-service" }]
-  },
-  "include": ["src/**/*"],
-  "exclude": ["node_modules", "dist"]
-}
